requirements.txt
aniso8601==9.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cryptography==36.0.0
Deprecated==1.2.13
Flask==2.0.2
Flask-Caching==1.10.1
Flask-HTTPAuth==4.5.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
pytz==2021.3
redis==4.0.2
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.27
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
python-dotenv==0.19.2
